+++ gen_data.py
@@ -1,4 +1,4 @@
-from .utils import gen_rand_sample_2d_data, encode
+from utils import gen_rand_sample_2d_data, encode
 import torch
 
 n = 4
@@ -7,8 +7,8 @@ n = 4
 with open("data.txt", "w") as f:
     # write lines of data to file
     for i in range(100000):
-        data = gen_rand_sample_2d_data(4)
-        p, v, k = data
-        k = torch.argmax(k.abs()).item()
-        tokens = encode(p, v, k)
+        data = gen_rand_sample_2d_data(4, 2)
+        v, p, dots = data
+        k = torch.argmax(dots.abs()).item()
+        tokens = encode(v, p, k)
         f.write(",".join([str(x) for x in tokens]) + "\n")
