+++ makemore.py
@@ -34,6 +34,8 @@ from torch.utils.data import Dataset
 from torch.utils.data.dataloader import DataLoader
 from torch.utils.tensorboard import SummaryWriter
 
+from model import Transformer
+
 # -----------------------------------------------------------------------------
 
 
@@ -54,185 +56,7 @@ def atomic_torch_save(dict, filename):
 
 
 # -----------------------------------------------------------------------------
-# Transformer Language Model (*exactly* as used in GPT-2)
-
-
-class NewGELU(nn.Module):
-    """
-    Implementation of the GELU activation function currently in Google BERT repo (identical to OpenAI GPT).
-    Reference: Gaussian Error Linear Units (GELU) paper: https://arxiv.org/abs/1606.08415
-    """
-
-    def forward(self, x):
-        return (
-            0.5
-            * x
-            * (
-                1.0
-                + torch.tanh(
-                    math.sqrt(2.0 / math.pi) * (x + 0.044715 * torch.pow(x, 3.0))
-                )
-            )
-        )
-
-
-class CausalSelfAttention(nn.Module):
-    """
-    A vanilla multi-head masked self-attention layer with a projection at the end.
-    It is possible to use torch.nn.MultiheadAttention here but I am including an
-    explicit implementation here to show that there is nothing too scary here.
-    """
-
-    def __init__(self, config):
-        super().__init__()
-        assert config.n_embd % config.n_head == 0
-        # key, query, value projections for all heads, but in a batch
-        self.c_attn = nn.Linear(config.n_embd, 3 * config.n_embd)
-        # output projection
-        self.c_proj = nn.Linear(config.n_embd, config.n_embd)
-        # causal mask to ensure that attention is only applied to the left in the input sequence
-        self.register_buffer(
-            "bias",
-            torch.tril(torch.ones(config.block_size, config.block_size)).view(
-                1, 1, config.block_size, config.block_size
-            ),
-        )
-        self.n_head = config.n_head
-        self.n_embd = config.n_embd
-
-    def forward(self, x):
-        B, T, C = (
-            x.size()
-        )  # batch size, sequence length, embedding dimensionality (n_embd)
-
-        # use einops to move around the head dimension
-        # calculate query, key, values for all heads in batch and move head forward to be the batch dim
-        # split the last dimension into 3 (q, k, v), then into (nh, hs)
-        q, k, v = einops.rearrange(
-            self.c_attn(x),
-            "B T (three embed) -> three B T embed",
-            three=3,
-            embed=self.n_embd,
-        )
-
-        k = einops.rearrange(
-            k, "B T (nh hs) -> B nh T hs", nh=self.n_head, hs=C // self.n_head
-        )
-        q = einops.rearrange(
-            q, "B T (nh hs) -> B nh T hs", nh=self.n_head, hs=C // self.n_head
-        )
-        v = einops.rearrange(
-            v, "B T (nh hs) -> B nh T hs", nh=self.n_head, hs=C // self.n_head
-        )
-
-        # causal self-attention; Self-attend: (B, nh, T, hs) x (B, nh, hs, T) -> (B, nh, T, T)
-
-        # swap the last two dimensions of k
-        k = einops.rearrange(k, "B nh T hs -> B nh hs T")
-
-        # T = S
-        att = einops.einsum(q, k, "B nh T hs, B nh hs S -> B nh T S") * (
-            1.0 / math.sqrt(k.size(-2))
-        )
-
-        # gpt style
-        att = att.masked_fill(self.bias[:, :, :T, :T] == 0, float("-inf"))
-
-        att = F.softmax(att, dim=-1)
-
-        # y = att @ v  # (B, nh, T, T) x (B, nh, T, hs) -> (B, nh, T, hs)
-        y = einops.einsum(att, v, "B nh T T, B nh T hs -> B nh T hs")
-
-        # transpose
-        y = einops.rearrange(y, "B nh T hs -> B T (nh hs)")
-
-        # output projection
-        y = self.c_proj(y)
-        return y
-
-
-class Block(nn.Module):
-    """an unassuming Transformer block"""
-
-    def __init__(self, config):
-        super().__init__()
-        self.ln_1 = nn.LayerNorm(config.n_embd)
-        self.attn = CausalSelfAttention(config)
-        self.ln_2 = nn.LayerNorm(config.n_embd)
-        self.mlp = nn.ModuleDict(
-            dict(
-                c_fc=nn.Linear(config.n_embd, 4 * config.n_embd),
-                c_proj=nn.Linear(4 * config.n_embd, config.n_embd),
-                act=NewGELU(),
-            )
-        )
-        m = self.mlp
-        self.mlpf = lambda x: m.c_proj(m.act(m.c_fc(x)))  # MLP forward
-
-    def forward(self, x):
-        x = x + self.attn(self.ln_1(x))
-        x = x + self.mlpf(self.ln_2(x))
-        return x
-
-
-class Transformer(nn.Module):
-    """Transformer Language Model, exactly as seen in GPT-2"""
-
-    def __init__(self, config):
-        super().__init__()
-        self.block_size = config.block_size
-
-        self.transformer = nn.ModuleDict(
-            dict(
-                wte=nn.Embedding(config.vocab_size, config.n_embd),
-                wpe=nn.Embedding(config.block_size, config.n_embd),
-                h=nn.ModuleList([Block(config) for _ in range(config.n_layer)]),
-                ln_f=nn.LayerNorm(config.n_embd),
-            )
-        )
-        self.lm_head = nn.Linear(config.n_embd, config.vocab_size, bias=False)
-
-        # report number of parameters (note we don't count the decoder parameters in lm_head)
-        n_params = sum(p.numel() for p in self.transformer.parameters())
-        print("number of parameters: %.2fM" % (n_params / 1e6,))
-
-    def get_block_size(self):
-        return self.block_size
-
-    def forward(self, idx, targets=None):
-        device = idx.device
-        b, t = idx.size()
-        assert t <= self.block_size, (
-            f"Cannot forward sequence of length {t}, block size is only {self.block_size}"
-        )
-        pos = torch.arange(0, t, dtype=torch.long, device=device).unsqueeze(
-            0
-        )  # shape (1, t)
-
-        # forward the GPT model itself
-        tok_emb = self.transformer.wte(idx)  # token embeddings of shape (b, t, n_embd)
-        pos_emb = self.transformer.wpe(
-            pos
-        )  # position embeddings of shape (1, t, n_embd)
-        x = tok_emb + pos_emb
-        for block in self.transformer.h:
-            x = block(x)
-        x = self.transformer.ln_f(x)
-        logits = self.lm_head(x)
-
-        # if we are given some desired targets also calculate the loss
-        loss = None
-        if targets is not None:
-            loss = F.cross_entropy(
-                logits.view(-1, logits.size(-1)), targets.view(-1), ignore_index=-1
-            )
-
-        return logits, loss
-
-
-# -----------------------------------------------------------------------------
 # helper functions for evaluating and sampling from the model
-
 
 @torch.no_grad()
 def generate(model, idx, max_new_tokens, temperature=1.0, do_sample=False, top_k=None):
