+++ utils.py
@@ -3,7 +3,9 @@ from functools import reduce
 from random import choice
 
 import torch
-from torch import argmax, cat, dot, empty, rand, stack
+from torch import argmax, dot, empty, rand, stack
+
+from torch import Tensor
 
 
 # rounds a tensor to n decimal places
@@ -12,21 +14,22 @@ def round_by(x, n):
     return x
 
 
-def gen_rand_sample_2d_data(n_points: int, round_by: int):
+def gen_rand_sample_2d_data(n_points: int, round_factor: int = 2):
     # uniform in [0.00,0.99]
     p = (0.00 - 0.99) * rand(n_points, 2) + 0.99
 
     # a linear dependence in [-0.99, 0.99]
     v = (-0.99 - 0.99) * rand(n_points) + 0.99
-
     # normalize to (0,1)
 
-    # round p and v to 2 decimal places
+    # truncate p and v to 2 decimal places
     # gives quantised numbers of 0-99
     # fits in a quantised 2^10-vector
+    p = torch.clamp(p, 0.00, 0.99)
+    v = torch.clamp(v, -0.99, 0.99)
 
-    p = round_by(p, round_by)
-    v = round_by(v, round_by)
+    p = torch.trunc(p * 100) / 100.0
+    v = torch.trunc(v * 100) / 100.0
 
     dots = empty(n_points)
 
@@ -65,7 +68,7 @@ def gen_rand_sample_2d_data(n_points: int, round_by: int):
 
     # max_index = argmax(dots.abs())
     # coeff_tensors.append(coeffs)
-    return p, v, dots
+    return v, p, dots
 
 
 # second experiment
@@ -81,9 +84,17 @@ END_BLOCK_TOKEN = 103
 CLASS_START = 104
 
 
-def encode(p, v, k):
+def encode(v, p, k):
     encoded = []
     n, _ = p.shape
+    for j in range(n):
+        vk = v[j]
+        if vk < 0:
+            vk *= -1
+            encoded.append(SIGN_TOKEN)
+        encoded.append(int(vk * 100) + 1)
+    encoded.append(END_BLOCK_TOKEN)
+    
     for j in range(n):
         x, y = p[j]
         if x < 0:
@@ -95,20 +106,14 @@ def encode(p, v, k):
             encoded.append(SIGN_TOKEN)
         encoded.append(int(y * 100) + 1)
     encoded.append(END_BLOCK_TOKEN)
-    for j in range(n):
-        vk = v[j]
-        if vk < 0:
-            vk *= -1
-            encoded.append(SIGN_TOKEN)
-        encoded.append(int(vk * 100) + 1)
-    encoded.append(END_BLOCK_TOKEN)
+    
     encoded.append(k + CLASS_START)
     return encoded
 
 
 def decode(encoded):
-    p = []
     v = []
+    p = []
     k = None
     is_negative = False  # Flag to track if the current number is negative
 
@@ -123,8 +128,21 @@ def decode(encoded):
             temp_block.append(num)
     blocks.append(temp_block)  # For the last block, which will be `k`
 
+    # Decode `v` block
+    v_block = blocks[0]
+    for num in v_block:
+        if num == SIGN_TOKEN:
+            is_negative = True
+            continue
+
+        val = (num - 1) / 100.0
+        if is_negative:
+            val *= -1
+            is_negative = False
+        v.append(val)
+    
     # Decode `p` block
-    p_block = blocks[0]
+    p_block = blocks[1]
     i = 0
     while i < len(p_block):
         if p_block[i] == SIGN_TOKEN:
@@ -142,19 +160,6 @@ def decode(encoded):
             p.append(temp_tuple + (val,))
         i += 1
 
-    # Decode `v` block
-    v_block = blocks[1]
-    for num in v_block:
-        if num == SIGN_TOKEN:
-            is_negative = True
-            continue
-
-        val = (num - 1) / 100.0
-        if is_negative:
-            val *= -1
-            is_negative = False
-        v.append(val)
-
     # Decode `k`
     k = blocks[2][0] - CLASS_START
 
@@ -162,10 +167,10 @@ def decode(encoded):
     v_tensor = torch.tensor(v, dtype=torch.float32)
     # k_tensor = torch.tensor([k], dtype=torch.int64)  # k is a single value, so wrap in a list
 
-    return p_tensor, v_tensor, k
+    return v_tensor, p_tensor, k
 
 
-def compute_max_dot(p, v):
+def compute_max_dot(v: Tensor, p: Tensor):
     # in general n points
     n_points = 4
     dots = empty(n_points)
@@ -212,11 +217,145 @@ def compute_max_dot(p, v):
 def decode_and_check(input):
     # there should be p, v and a k
     # p is 2 n ints, v is n ints, k is 1 int
-    p, v, k_out = decode(input)
-    dots, k_eval = compute_max_dot(p, v)
+    v, p, k_out = decode(input)
+    dots, k_eval = compute_max_dot(v, p)
     print(f"Predicted {k_out}, found {int(k_eval)} from computing dot of predicted")
     if abs(int(k_eval) - int(k_out)) < 0.001:
         return True
     else:
         return False
-    #
+
+
+# p is a list of tuples of coords
+# v is a list of values
+# v, p, k are python tensors
+def local_search(v, p, k):
+    # given a v, p and k,
+    # fix a v, and then try to find a better p
+    # by local search.
+    # k determines the pot product,
+
+    best_p = p.clone()
+    best_p_seen = [best_p.clone()]
+    best_dots, k_eval = compute_max_dot(v, best_p)
+    assert k_eval == k
+
+    best_dot_squared = best_dots[k_eval].abs() ** 2
+
+    max_num_steps = 20
+
+    # nelder mead coefficients
+    alpha = 1.0  # reflection
+    gamma = 2.0  # expansion
+    rho = 0.5  # contraction
+    sigma = 0.5  # shrink
+
+    step = 0
+    vertices = [best_p.clone()]
+    for _ in range(8):
+        vertices.append(best_p + 0.01 * (rand(best_p.shape) - 0.5))
+
+    while True:
+        step += 1
+
+        # order (ascending for minimization)
+        vertex_dots = [compute_max_dot(v, vtx) for vtx in vertices]
+        vertex_dot_values = [dots[k].abs() ** 2 for dots, k in vertex_dots]
+
+        sorted_indices = sorted(range(9), key=lambda i: vertex_dot_values[i])
+        vertices = [vertices[i] for i in sorted_indices]
+        vertex_dot_values = [vertex_dot_values[i] for i in sorted_indices]
+
+        # check if best is better than global best
+        if vertex_dot_values[0] < best_dot_squared:
+            best_p = vertices[0].clone()
+            best_dot_squared = vertex_dot_values[0]
+            best_p_seen.append(best_p.clone())
+
+        # terminate after max number
+        if step >= max_num_steps:
+            break
+
+        centroid = sum(vertices[:-1]) / 8.0
+
+        reflected = centroid + alpha * (centroid - vertices[-1])
+        reflected_dots, _ = compute_max_dot(v, reflected)
+        reflected_value = reflected_dots[k].abs() ** 2
+
+        # if reflected point is better than second worst but not better than best
+        if vertex_dot_values[0] <= reflected_value < vertex_dot_values[-2]:
+            vertices[-1] = reflected
+            continue
+
+        # expand
+        if reflected_value < vertex_dot_values[0]:
+            expanded = centroid + gamma * (centroid - vertices[-1])
+            expanded_dots, _ = compute_max_dot(v, expanded)
+            expanded_value = expanded_dots[k].abs() ** 2
+
+            if expanded_value < reflected_value:
+                vertices[-1] = expanded
+            else:
+                vertices[-1] = reflected
+            continue
+
+        # contract
+        if reflected_value >= vertex_dot_values[-2]:
+            if reflected_value < vertex_dot_values[-1]:
+                # outside contraction
+                contracted = centroid + rho * (reflected - centroid)
+                contracted_dots, _ = compute_max_dot(v, contracted)
+                contracted_value = contracted_dots[k].abs() ** 2
+
+                if contracted_value < reflected_value:
+                    vertices[-1] = contracted
+                    continue
+            else:
+                # inside contraction
+                contracted = centroid + rho * (vertices[-1] - centroid)
+                contracted_dots, _ = compute_max_dot(v, contracted)
+                contracted_value = contracted_dots[k].abs() ** 2
+
+                if contracted_value < vertex_dot_values[-1]:
+                    vertices[-1] = contracted
+                    continue
+
+        # shrint
+        best_vertex = vertices[0]
+        for i in range(1, 9):
+            vertices[i] = best_vertex + sigma * (vertices[i] - best_vertex)
+
+    # compute the dots of the best seen p's
+    best = [
+        (compute_max_dot(v, p_candidate)[0][k].abs() ** 2, p_candidate)
+        for p_candidate in best_p_seen
+    ]
+
+    # choose the smallest dot
+    # this may just be the original best p 
+    best_sorted = sorted(best, key=lambda pair: pair[0])
+
+    found_better = best_sorted[0][0] < best_dots[k].abs() ** 2
+
+    return found_better, best_sorted
+
+
+def test_local_search():
+    p = torch.rand(4, 2)
+    v = torch.rand(4)
+    dots, k_eval = compute_max_dot(v, p)
+    print(
+        f"Initial max dot at k={k_eval} with value {dots[k_eval].abs()} and p={p} and v={v}"
+    )
+
+    found_better, candidates = local_search(v, p, k_eval)
+
+    new_dot = candidates[0][0]
+    improved_p = candidates[0][1]
+
+    new_dots, new_k_eval = compute_max_dot(v, improved_p)
+    print(
+        f"After local search max dot at k {new_k_eval} with value {new_dots[new_k_eval].abs()} and p={improved_p} and v={v}, and found better is {found_better}"
+    )
+
+    assert new_dots[new_k_eval].abs() <= dots[k_eval].abs()
